Move WebErrorHandlerProvider::createNewApiProblem to WebErrorHandler class
This new class can more easily be used outside of Silex as it's not a ServiceProvider
